pyproject.toml
[project]
name = "lernix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "boto3>=1.42.15",
    "chromadb>=1.3.7",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.1",
    "fastapi>=0.127.0",
    "ipykernel>=7.1.0",
    "jq>=1.10.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.7",
    "pymysql>=1.1.2",
    "pypdf>=6.5.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
    "wikipedia>=1.4.0",
]
